+++ src/common/js/index.js
@@ -9,4 +9,6 @@ import './i18n.js';
 import './trace.js';
 //
 import './collections-get.js';
+import './event-emitter.js';
 import './functions.js';
+import './tenants-list.js';