+++ src/server/js/index.js
@@ -6,4 +6,3 @@ import '../../common/js/index.js';
 
 import './check_npms.js';
 import './collections.js';
-import './event-emitter.js';